requirements.txt
fastapi
uvicorn
python-dotenv
openai
pydantic
scikit-learn
jieba
requests
numpy
python-multipart
